adding namespace for search components
